requirements.txt
aiohttp
openai
pyrogram
python-dotenv